Add Missing IDEA project files. Split http backend and search crawler. Search items count optimization.

+++ Config/Configuration.go
@@ -32,7 +32,7 @@ type Configuration struct {
 	DhtConfig    DhtConfig
 	HttpConfig   HttpConfig
 	ScrapeConfig ScrapeConfig
-	ItemsPerPage int
+	ItemsPerPage uint64
 }
 
 func SetupConfiguration() *Configuration {
